Separate Form schema definitions from Form implementations
