+++ ui/src/components/forms/common/PriorityForm.schema.ts
@@ -1,0 +1,8 @@
+import { z } from 'zod';
+import { Priority } from '../../../types/enums';
+
+export const priorityFormSchema = z.object({
+  priority: z.nativeEnum(Priority),
+});
+
+export type PriorityFormState = z.infer<typeof priorityFormSchema>;
